chore: move "task" module into its own file
